backseat/autonomy/soi/SoiExecutive: Add cast to long on deadline creation.

+++ backseat/src/main/java/pt/lsts/autonomy/soi/SoiExecutive.java
@@ -120,7 +120,7 @@ public class SoiExecutive extends TimedFSM {
 	 */
 	public SoiExecutive() {
 		setPlanName(SOI_PLAN_ID);
-		setDeadline(new Date(System.currentTimeMillis() + timeout * 60 * 1000));
+		setDeadline(new Date(System.currentTimeMillis() + (long) timeout * 60 * 1000));
 		state = this::idleAtSurface;
 	}
